src/main/java/Ch16_The_Java_Memory_Model/C16_4_Thread_safe_lazy_initialization/SafeLazyInitialization.java
package Ch16_The_Java_Memory_Model.C16_4_Thread_safe_lazy_initialization;

import support.annotations.ThreadSafe;

/**
 * SafeLazyInitialization
 * <p/>
 * Thread-safe lazy initialization
 *
 * @author Brian Goetz and Tim Peierls
 */
@ThreadSafe
public class SafeLazyInitialization {
    private static Resource resource;

    public synchronized static Resource getInstance() {
        if (resource == null)
            resource = new Resource();
        return resource;
    }

    static class Resource {
    }
}
